LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/load64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "load64.wast:3"
PASS #3 Test that WebAssembly compilation fails
PASS #4 Test that WebAssembly compilation fails
PASS #5 Test that WebAssembly compilation fails
PASS #6 Test that WebAssembly compilation fails
PASS #7 Test that WebAssembly compilation fails
PASS #8 Test that WebAssembly compilation fails
PASS #9 Test that WebAssembly compilation fails
PASS #10 Test that WebAssembly compilation fails
PASS #11 Test that WebAssembly compilation fails
PASS #12 Test that WebAssembly compilation fails
PASS #13 Test that WebAssembly compilation fails
PASS #14 Test that WebAssembly compilation fails
PASS #15 Test that WebAssembly compilation fails
PASS #16 Test that WebAssembly compilation fails
PASS #17 Test that WebAssembly compilation fails
PASS #18 Test that WebAssembly compilation fails
PASS #19 Test that WebAssembly compilation fails
PASS #20 Test that WebAssembly compilation fails
PASS #21 Test that WebAssembly compilation fails
PASS #22 Test that WebAssembly compilation fails
PASS #23 Test that WebAssembly compilation fails
PASS #24 Test that WebAssembly compilation fails
PASS #25 Test that WebAssembly compilation fails
PASS #26 Test that WebAssembly compilation fails
PASS #27 Test that WebAssembly compilation fails
PASS #28 Test that WebAssembly compilation fails
PASS #29 Test that WebAssembly compilation fails
PASS #30 Test that WebAssembly compilation fails
PASS #31 Test that WebAssembly compilation fails
PASS #32 Test that WebAssembly compilation fails
PASS #33 Test that WebAssembly compilation fails
PASS #34 Test that WebAssembly compilation fails
PASS #35 Test that WebAssembly compilation fails
PASS #36 Test that WebAssembly compilation fails
PASS #37 Test that WebAssembly compilation fails
PASS #38 Test that WebAssembly compilation fails
PASS #39 Test that WebAssembly compilation fails
PASS #40 Test that WebAssembly compilation fails
PASS #41 Test that WebAssembly compilation fails
PASS #42 Test that WebAssembly compilation fails
PASS #43 Test that WebAssembly compilation fails
PASS #44 Test that WebAssembly compilation fails
PASS #45 Test that WebAssembly compilation fails
PASS #46 Test that WebAssembly compilation fails
PASS #47 Test that WebAssembly compilation fails
PASS #48 Test that WebAssembly compilation fails
PASS #49 Test that WebAssembly compilation fails
PASS #50 Test that WebAssembly compilation fails
PASS #51 Test that WebAssembly compilation fails
PASS #52 Test that WebAssembly compilation fails
PASS #53 Test that WebAssembly compilation fails
PASS #54 Test that WebAssembly compilation fails
PASS #55 Test that WebAssembly compilation fails
PASS #56 Test that WebAssembly compilation fails
PASS #57 Test that WebAssembly compilation fails
PASS #58 Test that WebAssembly compilation fails
PASS #59 Test that WebAssembly compilation fails
PASS #60 Test that WebAssembly compilation fails
PASS #61 Test that WebAssembly compilation fails
PASS #62 Reinitialize the default imports
FAIL #63 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 103: Memory64 is not enabled at {loc} expected true got false
FAIL #64 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #65 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #66 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:13:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #67 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:16:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #68 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #69 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #70 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #71 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #72 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #73 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:34:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #74 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #75 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:40:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #76 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:43:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #77 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:46:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #78 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:49:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #79 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #80 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:55:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #81 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:58:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #82 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #83 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #84 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:67:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #85 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:70:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #86 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:73:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #87 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:76:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #88 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:79:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #89 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:82:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #90 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:85:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #91 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:88:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #92 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:91:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #93 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:94:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #94 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:97:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #95 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:100:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #96 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:103:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #97 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:106:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #98 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:109:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #99 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:112:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #100 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:115:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
FAIL #101 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
load64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:118:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/load64.wast.js:297:3 expected true got false
PASS #102 Test that WebAssembly compilation fails
PASS #103 Test that WebAssembly compilation fails
PASS #104 Test that WebAssembly compilation fails
PASS #105 Test that WebAssembly compilation fails
PASS #106 Test that WebAssembly compilation fails
PASS #107 Test that WebAssembly compilation fails
PASS #108 Test that WebAssembly compilation fails
PASS #109 Test that WebAssembly compilation fails
PASS #110 Test that WebAssembly compilation fails
PASS #111 Test that WebAssembly compilation fails
PASS #112 Test that WebAssembly compilation fails
PASS #113 Test that WebAssembly compilation fails
PASS #114 Test that WebAssembly compilation fails
PASS #115 Test that WebAssembly compilation fails
PASS #116 Test that WebAssembly compilation fails
PASS #117 Test that WebAssembly compilation fails
PASS #118 Test that WebAssembly compilation fails
PASS #119 Test that WebAssembly compilation fails
PASS #120 Test that WebAssembly compilation fails
PASS #121 Test that WebAssembly compilation fails
PASS #122 Test that WebAssembly compilation fails
PASS #123 Test that WebAssembly compilation fails
PASS #124 Test that WebAssembly compilation fails
PASS #125 Test that WebAssembly compilation fails
PASS #126 Test that WebAssembly compilation fails
PASS #127 Test that WebAssembly compilation fails
PASS #128 Test that WebAssembly compilation fails
PASS #129 Test that WebAssembly compilation fails
PASS #130 Test that WebAssembly compilation fails
PASS #131 Test that WebAssembly compilation fails
PASS #132 Test that WebAssembly compilation fails
PASS #133 Test that WebAssembly compilation fails
PASS #134 Test that WebAssembly compilation fails
PASS #135 Test that WebAssembly compilation fails
PASS #136 Test that WebAssembly compilation fails
PASS #137 Test that WebAssembly compilation fails
PASS #138 Test that WebAssembly compilation fails
PASS #139 Test that WebAssembly compilation fails
PASS #140 Test that WebAssembly compilation fails
PASS #141 Test that WebAssembly compilation fails
PASS #142 Test that WebAssembly compilation fails
PASS #143 Test that WebAssembly compilation fails
PASS #144 Test that WebAssembly compilation fails
PASS #145 Test that WebAssembly compilation fails
PASS #146 Test that WebAssembly compilation fails
PASS #147 Test that WebAssembly compilation fails
PASS #148 Test that WebAssembly compilation fails
PASS #149 Test that WebAssembly compilation fails
PASS #150 Test that WebAssembly compilation fails
PASS #151 Test that WebAssembly compilation fails
PASS #152 Test that WebAssembly compilation fails
PASS #153 Test that WebAssembly compilation fails
PASS #154 Test that WebAssembly compilation fails
PASS #155 Test that WebAssembly compilation fails
PASS #156 Test that WebAssembly compilation fails
PASS #157 Test that WebAssembly compilation fails
PASS #158 Test that WebAssembly compilation fails
PASS #159 Test that WebAssembly compilation fails
PASS #160 Test that WebAssembly compilation fails
